+++ matchengine/me_message.h
@@ -18,7 +18,7 @@ enum {
     ORDER_EVENT_FINISH  = 3,
 };
 
-int push_balance_message(double t, uint32_t user_id, const char *asset, const char *business, mpd_t *change);
+int push_balance_message(double t, uint64_t user_id, const char *asset, const char *business, mpd_t *change);
 int push_order_message(uint32_t event, order_t *order, market_t *market);
 int push_deal_message(double t, const char *market, order_t *ask, order_t *bid, mpd_t *price, mpd_t *amount,
         mpd_t *ask_fee, mpd_t *bid_fee, int side, uint64_t id, const char *stock, const char *money);
@@ -27,4 +27,3 @@ bool is_message_block(void);
 sds message_status(sds reply);
 
 # endif
-
